Added Localization for better Translation-options

+++ OpenSSHALib/Models/SSHKey.cs
@@ -23,12 +23,8 @@ public class SshKey
         };
         readerProcess.Start();
         var outputOfProcess = readerProcess.StandardOutput.ReadToEnd().Split(' ');
-        // if (int.TryParse(outputOfProcess[0], out var parsed)) KeySize = parsed;
         Fingerprint = outputOfProcess[1];
         Comment = outputOfProcess[2];
-        // KeyType = outputOfProcess[3].Replace("(", "").Replace(")", "").Trim();
-
-        //
 
         var keyTypeText = outputOfProcess[3].Replace("(", "").Replace(")", "").Trim();
 
@@ -41,22 +37,15 @@ public class SshKey
             throw new ArgumentException($"{keyTypeText} is not a valid enum member of {typeof(KeyType)}");
         }
 
-        //
-
         if (IsPublicKey) PrivateKey = new SshKey(AbsoluteFilePath.Replace(".pub", ""));
     }
 
     public string AbsoluteFilePath { get; }
     private bool IsPublicKey => AbsoluteFilePath.EndsWith(".pub");
+    public string KeyTypeString => IsPublicKey ? "public" : "private";
     public string Filename { get; }
-
     public string Comment { get; private set; }
-
-    // public SshKeyType KeySize { get; private set; }
-    // public int KeySize { get; private set; }
-    public SshKeyType KeyType { get; private set; }
-
-    // public string KeyType { get; private set; }
+    public SshKeyType KeyType { get; private set; } = new(Enums.KeyType.RSA);
     public string Fingerprint { get; private set; }
     public SshKey? PrivateKey { get; }
 
@@ -79,6 +68,6 @@ public class SshKey
     public void DeleteKeys()
     {
         File.Delete(AbsoluteFilePath);
-        File.Delete(PrivateKey.AbsoluteFilePath);
+        if(PrivateKey is not null) File.Delete(PrivateKey.AbsoluteFilePath);
     }
 }
